HG-7 refactor: 크롤러의 download_image 메소드 BaseCrawler로 포함
